refactor: . move dateEditorLabelContent() to a separate file and rename it to editTaskLabelContent()
